myt/requirements.txt
asgiref==3.3.4
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.9
Django==3.2
django-spa==0.3.5
djangorestframework==3.12.4
idna==3.3
jsonfield==3.1.0
optional-django==0.1.0
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
urllib3==1.26.7
whitenoise==5.0.1
